+++ src/components/layout/RootLayout/components/A2hsSnackbar/style.ts
@@ -1,27 +1,22 @@
 import CustomButton from '@/components/CustomMui/CustomButton';
+import CustomSnackbar from '@/components/CustomMui/CustomSnackbar';
 import styled from '@emotion/styled';
-import { Snackbar as SnackbarBase } from '@mui/material';
 
-const Snackbar = styled(SnackbarBase)`
+const Snackbar = styled(CustomSnackbar)`
   position: fixed;
   bottom: 80px;
-  left: 50%;
   width: 100%;
   max-width: 768px;
   padding: 0 16px;
   touch-action: none;
   cursor: grab;
   user-select: none;
-  transform: translateX(-50%);
 
   & .MuiPaper-root {
     display: flex;
     justify-content: space-between;
     width: 100%;
     height: 42px;
-    padding: 0 16px;
-    color: #212121;
-    background-color: ${({ theme }) => theme.colors.white};
 
     & .MuiSnackbarContent-action {
       padding: 0;
